Refactored Circuit Breaker Config.
Turned Circuit Breaker Config into an interface and added it to the Annotation Config Factory.

In preperation for adding an alternative implementaion for Fault Tolerance 2.1
