+++ dev/com.ibm.ws.microprofile.faulttolerance.cdi/src/com/ibm/ws/microprofile/faulttolerance/cdi/FaultToleranceInterceptor.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017,2018 IBM Corporation and others.
+ * Copyright (c) 2017, 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -125,7 +125,7 @@ public class FaultToleranceInterceptor {
                     retry = new RetryConfig(targetClass, (Retry) annotation);
                     retry.validate();
                 } else if (annotation.annotationType().equals(CircuitBreaker.class)) {
-                    circuitBreaker = new CircuitBreakerConfig(targetClass, (CircuitBreaker) annotation);
+                    circuitBreaker = annotationConfigFactory.createCircuitBreakerConfig(targetClass, (CircuitBreaker) annotation);
                     circuitBreaker.validate();
                 } else if (annotation.annotationType().equals(Timeout.class)) {
                     timeout = new TimeoutConfig(targetClass, (Timeout) annotation);
@@ -153,7 +153,7 @@ public class FaultToleranceInterceptor {
                     retry = new RetryConfig(method, targetClass, (Retry) annotation);
                     retry.validate();
                 } else if (annotation.annotationType().equals(CircuitBreaker.class)) {
-                    circuitBreaker = new CircuitBreakerConfig(method, targetClass, (CircuitBreaker) annotation);
+                    circuitBreaker = annotationConfigFactory.createCircuitBreakerConfig(method, targetClass, (CircuitBreaker) annotation);
                     circuitBreaker.validate();
                 } else if (annotation.annotationType().equals(Timeout.class)) {
                     timeout = new TimeoutConfig(method, targetClass, (Timeout) annotation);
