+++ dev/com.ibm.ws.microprofile.faulttolerance.cdi/src/com/ibm/ws/microprofile/faulttolerance/cdi/config/impl/AnnotationConfigFactoryImpl20.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
+ * Copyright (c) 2018, 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -13,9 +13,11 @@ package com.ibm.ws.microprofile.faulttolerance.cdi.config.impl;
 import java.lang.reflect.Method;
 
 import org.eclipse.microprofile.faulttolerance.Asynchronous;
+import org.eclipse.microprofile.faulttolerance.CircuitBreaker;
 
 import com.ibm.ws.microprofile.faulttolerance.cdi.config.AnnotationConfigFactory;
 import com.ibm.ws.microprofile.faulttolerance.cdi.config.AsynchronousConfig;
+import com.ibm.ws.microprofile.faulttolerance.cdi.config.CircuitBreakerConfig;
 
 /**
  * Factory for creating annotation config for FT 2.0
@@ -34,6 +36,18 @@ public class AnnotationConfigFactoryImpl20 implements AnnotationConfigFactory {
     @Override
     public AsynchronousConfig createAsynchronousConfig(Class<?> annotatedClass, Asynchronous annotation) {
         return new AsynchronousConfigImpl20(annotatedClass, annotation);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public CircuitBreakerConfig createCircuitBreakerConfig(Method annotatedMethod, Class<?> annotatedClass, CircuitBreaker annotation) {
+        return new CircuitBreakerConfigImpl(annotatedMethod, annotatedClass, annotation);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public CircuitBreakerConfig createCircuitBreakerConfig(Class<?> annotatedClass, CircuitBreaker annotation) {
+        return new CircuitBreakerConfigImpl(annotatedClass, annotation);
     }
 
 }
